backend/requirements.txt
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
flask_socketio==5.3.6
flask-cors==4.0.0
Werkzeug==2.2.3
itsdangerous==2.1.2
pyjwt==2.8.0
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.15
pydantic==2.5.2
email-validator==2.1.0.post1
python-arango==7.5.0
pytest==7.4.0
waitress==2.1.2
langgraph-checkpoint-sqlite==2.0.5